Moved functions using pytest to test folder
