+++ test/integ/test_multi_statement.py
@@ -18,10 +18,7 @@ import snowflake.connector.cursor
 from snowflake.connector import ProgrammingError, errors
 
 try:
-    from snowflake.connector.test_util import (
-        _wait_until_query_success,
-        _wait_while_query_running,
-    )
+    from test.helpers import _wait_until_query_success, _wait_while_query_running
 except ModuleNotFoundError:
     pass
